refactor: Home Heading 분리
