+++ src/components/Home/Heading/styles.ts
@@ -1,0 +1,11 @@
+import styled from '@emotion/native';
+import { TextBold } from '~components/Common/Text';
+
+export const Heading = styled.View`
+  text-align: center;
+  margin-bottom: 32px;
+`;
+
+export const HeadingText = styled(TextBold)`
+  text-align: center;
+`;
