+++ src/screens/Home/index.tsx
@@ -5,6 +5,7 @@ import Avatar1 from '~assets/avatars/Avatar1.svg';
 import DogHand from '~assets/dogs/dog-hand.svg';
 import { ActionButton } from '~components/Common/ActionButton';
 import { Icon } from '~components/Common/Icons';
+import { Heading } from '~components/Home/Heading';
 import { WalkInfo } from '~components/Home/WalkInfo';
 import { WalkInfoFallback } from '~components/Home/WalkInfo/fallback';
 import { WalkInfoLoader } from '~components/Home/WalkInfo/loader';
@@ -21,10 +22,7 @@ export const HomeScreen = ({ navigation }: Props) => {
         <Avatar1 width={32} height={32} />
         <Icon.Bell onPress={() => navigation.navigate('Notification')} />
       </S.Header>
-      <S.Heading>
-        <S.HeadingText fontSize={24}>오늘은 아빠랑</S.HeadingText>
-        <S.HeadingText fontSize={24}>산책가는 날!</S.HeadingText>
-      </S.Heading>
+      <Heading />
       <DogHand />
       <ErrorBoundary FallbackComponent={WalkInfoFallback}>
         <Suspense fallback={<WalkInfoLoader />}>
